Handling existing addToCart Property

+++ hamrostore/addToCart.js
@@ -1,4 +1,7 @@
 import { getCartProductFromLS } from "./getCartProducts";
+import { updateCartValue } from "./updateCartValue";
+
+getCartProductFromLS();
 
 export const addToCart = (event, id, stock)=>{
 
@@ -13,11 +16,27 @@ export const addToCart = (event, id, stock)=>{
     // console.log(quantity, price);
 
     price = price.replace('रु','' );
+
+    let existingProd = arrLocalStorageProduct.find((curProd)=>curProd.id === id);
+
+    if (existingProd){
+      alert('Already Added')
+      return false;
+    }
+
+
+
     price = Number(price * quantity);
     quantity = Number(quantity);
 
     let updateCart = {id, quantity, price};
     arrLocalStorageProduct.push(updateCart);
     localStorage.setItem('cartProductLS', JSON.stringify(arrLocalStorageProduct));
+
+    //update the cart button value
+
+    updateCartValue(arrLocalStorageProduct);
+
+
 
 };
